% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gordon}
\alias{gordon}
\alias{gordon_3d}
\title{gordon atlas}
\format{
An object of class \code{brain_atlas} of length 4.

An object of class \code{ggseg3d_atlas} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 4 rows and 4 columns.
}
\usage{
gordon

gordon_3d
}
\description{
These parcels had highly homogenous RSFC patterns, 
indicating that they contained one unique RSFC signal; 
furthermore, the parcels were much more homogenous than a 
null model matched for parcel size when tested in two separate datasets. 
Several alternative parcellation schemes were tested this way, and no 
other parcellation was as homogenous as or had as large a difference compared 
with its null model.
annot files were acquired from (\href{https://github.com/faskowit/multiAtlasTT}{faskowit/multiAtlasTT})
}
\examples{
data(gordon)
data(gordon_3d)
}
\references{
Gordon, E. M., Laumann, T. O., Adeyemo, B., Huckins, J. F., 
   Kelley, W. M., & Petersen, S. E. (2014). Generation and evaluation of
   a cortical area parcellation from resting-state correlations. 
   Cerebral cortex, 26(1), 288-303. Chicago
   (\href{https://pubmed.ncbi.nlm.nih.gov/25316338/}{PubMed})
    
\itemize{
 \item{gordon}{ - gordon atlas}
 \item{gordon_3d}{ - gordon 3d mesh atlas}
}
}
\concept{ggseg_atlases ggseg3d_atlases}
\keyword{datasets}
