% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{gordon}
\alias{gordon}
\title{Gordon Resting-State Cortical Parcellation}
\usage{
gordon()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\description{
Brain atlas for the Gordon cortical parcellation with 329 regions
based on resting-state functional connectivity boundaries.
Contains 2D polygon geometry for \code{\link[ggseg:ggbrain]{ggseg::geom_brain()}}.
}
\examples{
gordon()
}
\references{
Gordon EM, Laumann TO, Adeyemo B, Huckins JF, Kelley WM,
Petersen SE (2016). Generation and Evaluation of a Cortical Area
Parcellation from Resting-State Correlations. \emph{Cerebral Cortex},
26(1):288-303.
\doi{10.1093/cercor/bhu239}
}
\concept{ggseg_atlases}
